[CP-1685] Rename folder from modals to update-os-modals
