+++ packages/app/src/overview/components/update-os-modals/checking-updates-modal/checking-updates-modal.interface.ts
