+++ packages/app/src/overview/components/update-os-modals/dev-update-modal/dev-update-modal.component.tsx
@@ -3,9 +3,9 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import { DevUpdateModalProps } from "App/overview/components/modals/dev-update-modal/dev-update-modal.inteface"
-import { CenteredText } from "App/overview/components/modals/dev-update-modal/dev-update-modal.styled"
-import { OSUpdateModal } from "App/overview/components/modals/os-update-modal"
+import { DevUpdateModalProps } from "App/overview/components/update-os-modals/dev-update-modal/dev-update-modal.inteface"
+import { CenteredText } from "App/overview/components/update-os-modals/dev-update-modal/dev-update-modal.styled"
+import { OSUpdateModal } from "App/overview/components/update-os-modals/os-update-modal"
 import { ModalMainText, RoundIconWrapper } from "App/ui/components/modal-dialog"
 import { IconType } from "App/__deprecated__/renderer/components/core/icon/icon-type"
 import Icon from "App/__deprecated__/renderer/components/core/icon/icon.component"
