+++ packages/app/src/overview/components/update-os-modals/downloading-update-interrupted-modal/downloading-update-interrupted-modal.interface.ts
