+++ packages/app/src/overview/components/update-os-modals/updating-spinner-modal/updating-spinner-modal.component.tsx
@@ -3,7 +3,7 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import { UpdatingSpinnerModalProps } from "App/overview/components/modals/updating-spinner-modal/updating-spinner-modal.interface"
+import { UpdatingSpinnerModalProps } from "App/overview/components/update-os-modals/updating-spinner-modal/updating-spinner-modal.interface"
 import LoaderModal from "App/ui/components/loader-modal/loader-modal.component"
 import { FunctionComponent } from "App/__deprecated__/renderer/types/function-component.interface"
 import { intl } from "App/__deprecated__/renderer/utils/intl"
