+++ packages/app/src/overview/components/update-os-modals/updating-spinner-modal/updating-spinner-modal.interface.ts
