+++ packages/app/src/overview/components/update-os-modals/updating-success-modal/updating-success-modal.interface.ts
